// src/features/cases-mangement/sessions/ExpertSessions/types/ExperstSessionType.ts
export interface ExpertSessionType {
  id: string;
  type?: string;                 
  expertReportNumber: string;      
  assignedAuthority: string;       
  assignmentDate: string;          
  expertOfficeName: string;        
  taskStartDate?: string;           
  subjectOfExpertise: string;      
  finalOpinion: string;            
  reportSubmissionDate: string;    
  objections?: string;              
  notes?: string;       
  status?: string;         
  

  
   

   
}